+++ src/Component/OpenApi2/Tests/fixtures/issue-770/expected/Normalizer/ProfileIpAclRulesNormalizer.php
@@ -38,10 +38,10 @@ class ProfileIpAclRulesNormalizer implements DenormalizerInterface, NormalizerIn
             return new Reference($data['$recursiveRef'], $context['document-origin']);
         }
         if (\array_key_exists('uplinkRateLimitingMbps', $data) && \is_int($data['uplinkRateLimitingMbps'])) {
-            $data['uplinkRateLimitingMbps'] = (double) $data['uplinkRateLimitingMbps'];
+            $data['uplinkRateLimitingMbps'] = (float) $data['uplinkRateLimitingMbps'];
         }
         if (\array_key_exists('downlinkRateLimitingMbps', $data) && \is_int($data['downlinkRateLimitingMbps'])) {
-            $data['downlinkRateLimitingMbps'] = (double) $data['downlinkRateLimitingMbps'];
+            $data['downlinkRateLimitingMbps'] = (float) $data['downlinkRateLimitingMbps'];
         }
         if (\array_key_exists('uplinkRateLimitingEnabled', $data) && \is_int($data['uplinkRateLimitingEnabled'])) {
             $data['uplinkRateLimitingEnabled'] = (bool) $data['uplinkRateLimitingEnabled'];
